Add new scripts that show my schedule
The main script is in python that reads an ics file. There are two capabilities this script offer. The first capability is to read then outputs the events inside the ics file into the console with pretty formatting. The second capability is to cache (create new ics file) from a file provided to the script by a specified start time and end time (or interval from now).